widgets: Add DBus and update widgets to use DBus for updates
Add a DBus loop (that gets threaded-off the main loop) to provide
background updates to the different widgets. This way, the widgets
don't have to poll for a status and we can save precious CPU cycles.

Note this still needs to be tested heavily, as in certain CPU
constrained situations this triggers an error with threading and
causes updates to not process from DBus.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>