settings: Hook up power and backlight controls to dbus
Make the necessary hooks into DBus to allow the shutdown, restart,
and backlight controls to start working. If backlight is not
available the control will not be present.

TODO: Need to check if we are allowed to restart or shutdown before
buttons are displayed and remove them from the settings if they
are not available.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>